+++ trunk/xorp/pim/pim_proto_assert.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/pim/pim_proto_assert.cc,v 1.10 2003/03/10 23:20:50 hodson Exp $"
+#ident "$XORP: xorp/pim/pim_proto_assert.cc,v 1.11 2003/05/21 05:32:54 pavlin Exp $"
 
 
 //
@@ -67,7 +67,7 @@ PimVif::pim_assert_recv(PimNbr *pim_nbr,
 			buffer_t *buffer)
 {
     int			rcvd_family;
-    uint32_t		group_masklen;
+    uint8_t		group_masklen;
     uint8_t		group_addr_reserved_flags;
     IPvX		assert_source_addr(family());
     IPvX		assert_group_addr(family());
@@ -134,7 +134,7 @@ PimVif::pim_assert_process(PimNbr *pim_nbr,
 			   const IPvX& src, const IPvX& dst,
 			   const IPvX& assert_source_addr,
 			   const IPvX& assert_group_addr,
-			   uint32_t group_masklen, AssertMetric *assert_metric)
+			   uint8_t group_masklen, AssertMetric *assert_metric)
 {
     PimMre	*pim_mre_sg, *pim_mre_wc;
     int ret_value;
@@ -351,7 +351,7 @@ PimVif::pim_assert_send(const IPvX& assert_source_addr,
 {
     buffer_t *buffer = buffer_send_prepare();
     uint8_t group_addr_reserved_flags = 0;
-    uint32_t group_masklen = IPvX::addr_bitlen(family());
+    uint8_t group_masklen = IPvX::addr_bitlen(family());
     
     if (rpt_bit)
 	metric_preference |= PIM_ASSERT_RPT_BIT;
